updated card views to look more like deck views

+++ api/views/card_views.py
@@ -1,7 +1,14 @@
-from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly
+from rest_framework.exceptions import PermissionDenied
+from rest_framework import generics, status
+from django.shortcuts import get_object_or_404
+from django.contrib.auth import get_user, authenticate, login, logout
+from django.middleware.csrf import get_token
+
+
+from django.shortcuts import render
 from django.shortcuts import get_object_or_404
 
 from ..models.card import Card
@@ -9,34 +16,45 @@ from ..serializers import CardSerializer
 
 # Create your views here.
 class Cards(APIView):
+    permission_classes=(IsAuthenticated,)
     def get(self, request):
         """Index Request"""
-        print(request)
-        cards = Card.objects.all()[:10]
+        # print(request)
+        cards = Card.objects.filter(owner=request.user.id)
+        # [:10] -- this could be a potential way to put a limit on how many cards to create!
         data = CardSerializer(cards, many=True).data
-        return Response(data)
+        return Response({ 'cards': data })
 
     serializer_class = CardSerializer
     def post(self, request):
         """Post request"""
-        print(request.data)
+        # print(request.data)
+        request.data['card']['owner'] = request.user.id
         card = CardSerializer(data=request.data['card'])
         if card.is_valid():
             card.save()
-            return Response(card.data, status=status.HTTP_201_CREATED)
+            return Response({ 'card': card.data }, status=status.HTTP_201_CREATED)
         else:
             return Response(card.errors, status=status.HTTP_400_BAD_REQUEST)
 
 class CardDetail(APIView):
+    permission_classes=(IsAuthenticated,)
     def get(self, request, pk):
         """Show request"""
         card = get_object_or_404(Card, pk=pk)
+        if not request.user.id == card.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this card')
         data = CardSerializer(card).data
-        return Response(data)
+        return Response({ 'card': data })
 
-    def patch(self, request, pk):
+    def partial_update(self, request, pk):
         """Update Request"""
         card = get_object_or_404(Card, pk=pk)
+        if request.data['card'].get('owner', False):
+            del request.data['card']['owner']
+        if not request.user.id == deck_instance.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this deck')
+        request.data['card']['owner'] = request.user.id
         ms = CardSerializer(card, data=request.data['card'], partial=True)
         if ms.is_valid():
             ms.save()
@@ -46,5 +64,7 @@ class CardDetail(APIView):
     def delete(self, request, pk):
         """Delete Request"""
         card = get_object_or_404(Card, pk=pk)
+        if not request.user.id == deck.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this card')
         card.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
